extract thread worker classes (#199)
* Pulled worker code into new class for concurrency manager. Unit tests passing

* Request rate worker class. 2 unit tests disabled. The rest are passing

* Remove moved code from LoadManager

* Created base worker class. Move common functions and struct defines into it

* Pushed more common variables to base class

* shared pointer for workers

* push threadconfig into worker classes. Remove circular dependencies

* Cleanup fixmes and comments

* Clean up and document member variables, and add TODOs

* Rename worker to loadworker

* Created interfaces for workers. Mocked out worker and revived test that relied on that

* Revive final test

* Tweak cons/destructors
