+++ src/c++/perf_analyzer/CMakeLists.txt
@@ -43,6 +43,9 @@ set(
   data_loader.cc
   concurrency_manager.cc
   request_rate_manager.cc
+  load_worker.cc
+  concurrency_worker.cc
+  request_rate_worker.cc
   custom_load_manager.cc
   inference_profiler.cc
   report_writer.cc
@@ -61,6 +64,9 @@ set(
   concurrency_manager.h
   request_rate_manager.h
   custom_load_manager.h
+  load_worker.h
+  request_rate_worker.h
+  concurrency_worker.h
   inference_profiler.h
   report_writer.h
   mpi_utils.h
