+++ src/c++/perf_analyzer/concurrency_manager.cc
@@ -25,7 +25,6 @@
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include "concurrency_manager.h"
-
 #include <queue>
 
 namespace triton { namespace perfanalyzer {
@@ -117,23 +116,22 @@ void
 ConcurrencyManager::ReconfigThreads(const size_t concurrent_request_count)
 {
   // Always prefer to create new threads if the maximum limit has not been met
+  //
+  // While operating in synchronous mode, each context can send only one
+  // request at a time, hence the number of worker threads should be equal to
+  // the requested concurrency levels.
+  //
   while ((concurrent_request_count > threads_.size()) &&
          (threads_.size() < max_threads_)) {
     // Launch new thread for inferencing
     threads_stat_.emplace_back(new ThreadStat());
-    threads_config_.emplace_back(new ThreadConfig(threads_config_.size()));
+    threads_config_.emplace_back(
+        new ConcurrencyWorker::ThreadConfig(threads_config_.size()));
 
-    // Worker maintains concurrency in different ways.
-    // For sequence models, multiple contexts must be created for multiple
-    // concurrent sequences.
-    // For non-sequence models, one context can send out multiple requests
-    // at the same time. Thus it uses one single context as every infer context
-    // creates a worker thread implicitly.
-    // While operating in synchronous mode, each context can send only one
-    // request at a time, hence the number of worker threads should be equal to
-    // the requested concurrency levels.
+    auto worker = MakeWorker();
+
     threads_.emplace_back(
-        &ConcurrencyManager::Infer, this, threads_stat_.back(),
+        &IConcurrencyWorker::Infer, worker, threads_stat_.back(),
         threads_config_.back());
   }
 
@@ -168,388 +166,16 @@ ConcurrencyManager::ResumeSequenceWorkers()
   wake_signal_.notify_all();
 }
 
-
-// Function for worker threads.
-// If the model is non-sequence model, each worker uses only one context
-// to maintain concurrency assigned to worker.
-// If the model is sequence model, each worker has to use multiples contexts
-// to maintain (sequence) concurrency assigned to worker.
-void
-ConcurrencyManager::Infer(
-    std::shared_ptr<ConcurrencyManager::ThreadStat> thread_stat,
-    std::shared_ptr<ConcurrencyManager::ThreadConfig> thread_config)
+std::shared_ptr<IConcurrencyWorker>
+ConcurrencyManager::MakeWorker()
 {
-  std::vector<std::unique_ptr<InferContext>> ctxs;
-  uint32_t seq_stat_index = 0, ctx_id = 0;
-  std::queue<int> free_ctx_ids;
-
-  // Reserve the vectors in case of sequence models. In non-sequence or
-  // synchronous mode only one context will be opened hence no need of
-  // reserving.
-  if (on_sequence_model_ && async_) {
-    thread_stat->contexts_stat_.reserve(max_concurrency_);
-    ctxs.reserve(max_concurrency_);
-  }
-
-  // Variable used to signal request completion
-  bool notified = false;
-  std::mutex cb_mtx;
-  std::condition_variable cb_cv;
-
-  std::atomic<int> total_ongoing_requests(0);
-  uint64_t request_id = 0;
-
-  // request_id to start timestamp map
-  std::map<std::string, AsyncRequestProperties> async_req_map;
-
-  // Callback function for handling asynchronous requests
-  const auto callback_func = [&](cb::InferResult* result) {
-    uint32_t ctx_id = 0;
-    std::shared_ptr<cb::InferResult> result_ptr(result);
-    if (thread_stat->cb_status_.IsOk()) {
-      // Add the request timestamp to thread Timestamp vector with
-      // proper locking
-      std::lock_guard<std::mutex> lock(thread_stat->mu_);
-      thread_stat->cb_status_ = result_ptr->RequestStatus();
-      if (thread_stat->cb_status_.IsOk()) {
-        std::chrono::time_point<std::chrono::system_clock> end_time_async;
-        end_time_async = std::chrono::system_clock::now();
-        std::string request_id;
-        thread_stat->cb_status_ = result_ptr->Id(&request_id);
-        const auto& it = async_req_map.find(request_id);
-        if (it != async_req_map.end()) {
-          thread_stat->request_timestamps_.emplace_back(std::make_tuple(
-              it->second.start_time_, end_time_async, it->second.sequence_end_,
-              false /* delayed */));
-          ctx_id = it->second.ctx_id_;
-          ctxs[ctx_id]->infer_backend_->ClientInferStat(
-              &(thread_stat->contexts_stat_[ctx_id]));
-          thread_stat->cb_status_ = ValidateOutputs(*ctxs[ctx_id], result);
-          async_req_map.erase(request_id);
-        }
-      }
-    }
-    // avoid competition over 'cb_mtx'
-    {
-      std::lock_guard<std::mutex> lk(cb_mtx);
-      free_ctx_ids.push(ctx_id);
-      notified = true;
-    }
-
-    total_ongoing_requests--;
-
-    cb_cv.notify_all();
-  };
-
-  // Specify the function as lambda here to work around the possible callback
-  // lifecycle issue when making this a class member function.
-  // Note that 'free_ctx_ids' must be reconstruct after the call because
-  // this function doesn't utilize 'free_ctx_ids' in the same way as in main
-  // loop
-  const auto complete_ongoing_sequence_func = [&]() {
-    if (!on_sequence_model_) {
-      return cb::Error::Success;
-    }
-    size_t offset = 0;
-    for (size_t i = 0; i < thread_config->thread_id_; i++) {
-      offset += threads_config_[i]->concurrency_;
-    }
-
-    for (size_t ctx_id = 0; ctx_id < ctxs.size(); ++ctx_id) {
-      size_t seq_stat_index = offset + ctx_id;
-
-      std::lock_guard<std::mutex> guard(sequence_stat_[seq_stat_index]->mtx_);
-      // Complete the sequence if there are remaining queries
-      while (sequence_stat_[seq_stat_index]->remaining_queries_ != 0) {
-        sequence_stat_[seq_stat_index]->remaining_queries_ = 1;
-        SetInferSequenceOptions(seq_stat_index, ctxs[ctx_id]->options_);
-
-        // Update the inputs if required
-        if (using_json_data_) {
-          int step_id = data_loader_->GetTotalSteps(
-                            sequence_stat_[seq_stat_index]->data_stream_id_) -
-                        sequence_stat_[seq_stat_index]->remaining_queries_;
-
-          RETURN_IF_ERROR(UpdateInputs(
-              ctxs[ctx_id]->inputs_, ctxs[ctx_id]->valid_inputs_,
-              sequence_stat_[seq_stat_index]->data_stream_id_, step_id));
-          RETURN_IF_ERROR(UpdateValidationOutputs(
-              ctxs[ctx_id]->outputs_,
-              sequence_stat_[seq_stat_index]->data_stream_id_, step_id,
-              ctxs[ctx_id]->expected_outputs_));
-        }
-        sequence_stat_[seq_stat_index]->remaining_queries_--;
-
-        if (async_) {
-          ctxs[ctx_id]->options_->request_id_ = "0";
-          if (streaming_) {
-            RETURN_IF_ERROR(ctxs[ctx_id]->infer_backend_->AsyncStreamInfer(
-                *(ctxs[ctx_id]->options_), ctxs[ctx_id]->inputs_,
-                ctxs[ctx_id]->outputs_));
-          } else {
-            RETURN_IF_ERROR(ctxs[ctx_id]->infer_backend_->AsyncInfer(
-                callback_func, *(ctxs[ctx_id]->options_), ctxs[ctx_id]->inputs_,
-                ctxs[ctx_id]->outputs_));
-          }
-          total_ongoing_requests++;
-        } else {
-          cb::InferResult* results = nullptr;
-          auto err = ctxs[ctx_id]->infer_backend_->Infer(
-              &results, *(ctxs[ctx_id]->options_), ctxs[ctx_id]->inputs_,
-              ctxs[ctx_id]->outputs_);
-          if (results != nullptr) {
-            delete results;
-          }
-          RETURN_IF_ERROR(err);
-        }
-      }
-    }
-    return cb::Error::Success;
-  };
-
-  // run inferencing until receiving exit signal to maintain server load.
-  do {
-    if (on_sequence_model_) {
-      if (!execute_) {
-        // Ensures the clean exit of the sequences
-        auto status = complete_ongoing_sequence_func();
-        if (thread_stat->status_.IsOk()) {
-          thread_stat->status_ = status;
-        }
-        while (total_ongoing_requests != 0) {
-          std::this_thread::sleep_for(std::chrono::milliseconds(500));
-        }
-        // Make sure all threads are in sync with the client's stats
-        //
-        for (size_t i = 0; i < ctxs.size(); ++i) {
-          ctxs[i]->infer_backend_->ClientInferStat(
-              &(thread_stat->contexts_stat_[i]));
-        }
-        // Reconstruct 'free_ctx_ids' because complete_ongoing_sequence_func()
-        // has destructive side affects
-        free_ctx_ids = std::queue<int>();
-        for (size_t i = 0; i < ctxs.size(); ++i) {
-          free_ctx_ids.push(i);
-        }
-        // Wait if no request should be sent and it is not exiting
-        thread_config->is_paused_ = true;
-        std::unique_lock<std::mutex> lock(wake_mutex_);
-        wake_signal_.wait(lock, [this]() { return early_exit || execute_; });
-      }
-    }
-
-    thread_config->is_paused_ = false;
-
-    // Only interact with synchronous mechanism if the worker should wait
-    if (thread_config->concurrency_ == 0) {
-      // Wait if no request should be sent and it is not exiting
-      std::unique_lock<std::mutex> lock(wake_mutex_);
-      wake_signal_.wait(lock, [&thread_config]() {
-        return early_exit || (thread_config->concurrency_ > 0);
-      });
-      // Stop executing if concurrency is 0 and early exit is requested
-      if (early_exit && thread_config->concurrency_ == 0) {
-        break;
-      }
-    }
-
-    size_t num_reqs = thread_config->concurrency_;
-
-    // If the model is non-sequence model, use one InferContext to maintain
-    // concurrency for this thread.
-    size_t active_ctx_cnt = on_sequence_model_ ? num_reqs : 1;
-
-    while (active_ctx_cnt > ctxs.size()) {
-      {
-        std::lock_guard<std::mutex> lock(cb_mtx);
-        free_ctx_ids.push(ctxs.size());
-      }
-      ctxs.emplace_back(new InferContext());
-      thread_stat->status_ =
-          factory_->CreateClientBackend(&(ctxs.back()->infer_backend_));
-      ctxs.back()->options_.reset(new cb::InferOptions(parser_->ModelName()));
-      ctxs.back()->options_->model_version_ = parser_->ModelVersion();
-      ctxs.back()->options_->model_signature_name_ =
-          parser_->ModelSignatureName();
-      thread_stat->contexts_stat_.emplace_back();
-      if (shared_memory_type_ == SharedMemoryType::NO_SHARED_MEMORY) {
-        thread_stat->status_ = PrepareInfer(ctxs.back().get());
-      } else {
-        thread_stat->status_ = PrepareSharedMemoryInfer(ctxs.back().get());
-      }
-      if (!thread_stat->status_.IsOk()) {
-        return;
-      }
-      if (streaming_) {
-        // Decoupled models should not collect client side statistics
-        thread_stat->status_ = ctxs.back()->infer_backend_->StartStream(
-            callback_func, (!parser_->IsDecoupled()));
-        if (!thread_stat->status_.IsOk()) {
-          return;
-        }
-      }
-    }
-
-    // Create async requests such that the number of ongoing requests
-    // matches the concurrency level
-    // Non-sequence model is 'num_reqs' * 1 ctx
-    // Sequence model is 1 request of 1 sequence * 'active_ctx_cnt' ctxs
-    while (total_ongoing_requests < (int)num_reqs && early_exit == false) {
-      // Update the inputs if required for non-sequence
-      if (using_json_data_ && (!on_sequence_model_)) {
-        int step_id = (thread_config->non_sequence_data_step_id_ %
-                       data_loader_->GetTotalStepsNonSequence()) *
-                      batch_size_;
-        thread_config->non_sequence_data_step_id_ += active_threads_;
-        // There will be only one ctx in non-sequence case
-        thread_stat->status_ = UpdateInputs(
-            ctxs[ctx_id]->inputs_, ctxs[ctx_id]->valid_inputs_, 0, step_id);
-        if (thread_stat->status_.IsOk()) {
-          thread_stat->status_ = UpdateValidationOutputs(
-              ctxs[ctx_id]->outputs_, 0, step_id,
-              ctxs[ctx_id]->expected_outputs_);
-        }
-        if (!thread_stat->status_.IsOk()) {
-          return;
-        }
-      }
-
-      if (on_sequence_model_) {
-        size_t offset = 0;
-        for (size_t i = 0; i < thread_config->thread_id_; i++) {
-          offset += threads_config_[i]->concurrency_;
-        }
-
-        // Find the next available context id to use for this request
-        {
-          std::lock_guard<std::mutex> lk(cb_mtx);
-          ctx_id = free_ctx_ids.front();
-          free_ctx_ids.pop();
-        }
-        seq_stat_index = offset + ctx_id;
-
-        {
-          std::lock_guard<std::mutex> guard(
-              sequence_stat_[seq_stat_index]->mtx_);
-          SetInferSequenceOptions(seq_stat_index, ctxs[ctx_id]->options_);
-
-          // Update the inputs if required
-          if (using_json_data_) {
-            int step_id = data_loader_->GetTotalSteps(
-                              sequence_stat_[seq_stat_index]->data_stream_id_) -
-                          sequence_stat_[seq_stat_index]->remaining_queries_;
-
-            thread_stat->status_ = UpdateInputs(
-                ctxs[ctx_id]->inputs_, ctxs[ctx_id]->valid_inputs_,
-                sequence_stat_[seq_stat_index]->data_stream_id_, step_id);
-            if (thread_stat->status_.IsOk()) {
-              thread_stat->status_ = UpdateValidationOutputs(
-                  ctxs[ctx_id]->outputs_,
-                  sequence_stat_[seq_stat_index]->data_stream_id_, step_id,
-                  ctxs[ctx_id]->expected_outputs_);
-            }
-            if (!thread_stat->status_.IsOk()) {
-              return;
-            }
-          }
-          sequence_stat_[seq_stat_index]->remaining_queries_--;
-        }
-      }
-      if (async_) {
-        ctxs[ctx_id]->options_->request_id_ = std::to_string(request_id++);
-        {
-          std::lock_guard<std::mutex> lock(thread_stat->mu_);
-          auto it = async_req_map
-                        .emplace(
-                            ctxs[ctx_id]->options_->request_id_,
-                            AsyncRequestProperties())
-                        .first;
-          it->second.start_time_ = std::chrono::system_clock::now();
-          it->second.ctx_id_ = ctx_id;
-          it->second.sequence_end_ = ctxs[ctx_id]->options_->sequence_end_;
-        }
-        if (streaming_) {
-          thread_stat->status_ = ctxs[ctx_id]->infer_backend_->AsyncStreamInfer(
-              *(ctxs[ctx_id]->options_), ctxs[ctx_id]->valid_inputs_,
-              ctxs[ctx_id]->outputs_);
-        } else {
-          thread_stat->status_ = ctxs[ctx_id]->infer_backend_->AsyncInfer(
-              callback_func, *(ctxs[ctx_id]->options_),
-              ctxs[ctx_id]->valid_inputs_, ctxs[ctx_id]->outputs_);
-        }
-        if (!thread_stat->status_.IsOk()) {
-          return;
-        }
-      } else {
-        std::chrono::time_point<std::chrono::system_clock> start_time_sync,
-            end_time_sync;
-        start_time_sync = std::chrono::system_clock::now();
-        cb::InferResult* results = nullptr;
-        thread_stat->status_ = ctxs[ctx_id]->infer_backend_->Infer(
-            &results, *(ctxs[ctx_id]->options_), ctxs[ctx_id]->valid_inputs_,
-            ctxs[ctx_id]->outputs_);
-        if (results != nullptr) {
-          if (thread_stat->status_.IsOk()) {
-            thread_stat->status_ = ValidateOutputs(*ctxs[ctx_id], results);
-          }
-          delete results;
-        }
-        if (!thread_stat->status_.IsOk()) {
-          return;
-        }
-        end_time_sync = std::chrono::system_clock::now();
-        {
-          // Add the request timestamp to thread Timestamp vector with proper
-          // locking
-          std::lock_guard<std::mutex> lock(thread_stat->mu_);
-          thread_stat->request_timestamps_.emplace_back(std::make_tuple(
-              start_time_sync, end_time_sync,
-              ctxs[ctx_id]->options_->sequence_end_, false /* delayed */));
-          thread_stat->status_ = ctxs[ctx_id]->infer_backend_->ClientInferStat(
-              &(thread_stat->contexts_stat_[ctx_id]));
-          if (!thread_stat->status_.IsOk()) {
-            return;
-          }
-        }
-        {
-          std::lock_guard<std::mutex> lock(cb_mtx);
-          free_ctx_ids.push(ctx_id);
-        }
-      }
-      total_ongoing_requests++;
-    }
-
-    if (async_) {
-      {
-        // If async, then wait for signal from callback.
-        std::unique_lock<std::mutex> lk(cb_mtx);
-        cb_cv.wait(lk, [&notified] {
-          if (notified) {
-            notified = false;
-            return true;
-          }
-          return false;
-        });
-      }
-    } else {
-      // If synchronous, then all the requests have already been completed.
-      total_ongoing_requests = 0;
-    }
-
-    if (early_exit || (!thread_stat->cb_status_.IsOk())) {
-      // Wait for all callbacks to complete.
-      // Loop to ensure all the inflight requests have been completed.
-      auto status = complete_ongoing_sequence_func();
-      if (thread_stat->status_.IsOk()) {
-        thread_stat->status_ = status;
-      }
-      while (total_ongoing_requests != 0) {
-        std::this_thread::sleep_for(std::chrono::milliseconds(500));
-      }
-      // end loop
-      break;
-    }
-  } while (true);
+  return std::make_shared<ConcurrencyWorker>(
+      parser_, data_loader_, backend_->Kind(), factory_, sequence_length_,
+      start_sequence_id_, sequence_id_range_, on_sequence_model_, async_,
+      max_concurrency_, using_json_data_, streaming_, shared_memory_type_,
+      batch_size_, threads_config_, sequence_stat_, shared_memory_regions_,
+      wake_signal_, wake_mutex_, active_threads_, execute_, curr_seq_id_,
+      distribution_);
 }
 
 }}  // namespace triton::perfanalyzer
